Q1 and Q2 Wrangled
Q1: What time of the year are cats and dogs more likely to get adopted?
and
Q2: What influences the number of cats and dogs in a shelter over time?

Columns are set up and some analysis has been done for.

Still have to create a dictionary for dogs and then separate columns based on dog ages.